fix(payment): Handle deleted customers
