+++ spec/services/payment_provider_customers/stripe_service_spec.rb
@@ -449,63 +449,6 @@ RSpec.describe PaymentProviderCustomers::StripeService, type: :service do
     end
   end
 
-  describe "#check_payment_method" do
-    let(:payment_method_id) { "card_12345" }
-
-    let(:stripe_customer) do
-      create(
-        :stripe_customer,
-        customer:,
-        provider_customer_id: "cus_123456",
-        payment_method_id:
-      )
-    end
-
-    let(:payment_method) { Stripe::PaymentMethod.new(id: payment_method_id) }
-
-    let(:stripe_api_customer) { instance_double(Stripe::Customer) }
-
-    before do
-      allow(Stripe::Customer).to receive(:new)
-        .and_return(stripe_api_customer)
-    end
-
-    it "checks for the existance of the payment method" do
-      allow(stripe_api_customer)
-        .to receive(:retrieve_payment_method)
-        .and_return(payment_method)
-
-      result = stripe_service.check_payment_method(payment_method_id)
-
-      aggregate_failures do
-        expect(result).to be_success
-        expect(result.payment_method.id).to eq(payment_method_id)
-
-        expect(Stripe::Customer).to have_received(:new)
-        expect(stripe_api_customer).to have_received(:retrieve_payment_method)
-      end
-    end
-
-    context "when payment method is not found on stripe" do
-      before do
-        allow(stripe_api_customer)
-          .to receive(:retrieve_payment_method)
-          .and_raise(::Stripe::InvalidRequestError.new("error", {}))
-      end
-
-      it "returns a failed result" do
-        result = stripe_service.check_payment_method(payment_method_id)
-
-        aggregate_failures do
-          expect(result).not_to be_success
-
-          expect(Stripe::Customer).to have_received(:new)
-          expect(stripe_api_customer).to have_received(:retrieve_payment_method)
-        end
-      end
-    end
-  end
-
   describe "#generate_checkout_url" do
     it "delivers a webhook with checkout url" do
       allow(::Stripe::Checkout::Session).to receive(:create)
